requirements.txt
pyserial
# tqdm # optional